+++ aoe2/aoe2_game_overlay.py
@@ -359,6 +359,8 @@ class AoE2GameOverlay(RTSGameOverlay):
             resources_line += spacing + '@' + images.villager + '@ ' + (
                 str(target_villager) if (target_villager >= 0) else ' ')
             resources_line += spacing + '@' + self.get_age_image(selected_step['age'])
+            if 'time' in selected_step:  # add time if indicated
+                resources_line += '@' + spacing + '@' + self.settings.images.time + '@' + selected_step['time']
 
             self.build_order_resources.add_row_from_picture_line(parent=self, line=str(resources_line))